Seperate out calendar components and UI Changes
- Updated layout
- Updated body bg color in `_document.tsx`
- Built Calendar and CalendarCell components
