+++ pages/_document.tsx
@@ -4,7 +4,7 @@ export default class MyDocument extends Document {
     return (
       <Html lang="en">
         <Head></Head>
-        <body>
+        <body className="bg-white dark:bg-dark-background">
           <Main />
           <NextScript />
         </body>
